change precision

+++ server/src/database/entities.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float, TIMESTAMP
+from sqlalchemy import Column, String, Numeric, TIMESTAMP
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 # from sqlalchemy.ext.declarative import declarative_base
@@ -38,9 +38,9 @@ class LineChart(Base):
     __tablename__ = "line_chart"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    temperature = Column(Float, nullable=False)
-    pressure = Column(Float, nullable=False)
-    humidity = Column(Float, nullable=False)
+    temperature = Column(Numeric(6,2), nullable=False)
+    pressure = Column(Numeric(6,2), nullable=False)
+    humidity = Column(Numeric(6,2), nullable=False)
     time_stamp = Column(TIMESTAMP, default=datetime.utcnow, index=True, nullable=False)
 
 
@@ -48,9 +48,9 @@ class BarChart(Base):
     __tablename__ = "bar_chart"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    steel = Column(Float, nullable=False)
-    lubricant = Column(Float, nullable=False)
-    anti_corrosion_Coating = Column(Float, nullable=False)
+    steel = Column(Numeric(6,2), nullable=False)
+    lubricant = Column(Numeric(6,2), nullable=False)
+    anti_corrosion_Coating = Column(Numeric(6,2), nullable=False)
     time_stamp = Column(TIMESTAMP, default=datetime.utcnow, index=True, nullable=False)
